integration_tests/requirements.txt
pytest==7.1.3
requests==2.28.1
grpcio==1.48.2
protobuf==4.21.6